requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-chroma>=0.1.0
langchain-text-splitters>=0.0.1

# Vector store and embeddings
chromadb>=0.4.22
openai>=1.10.0

# Document processing
pypdf>=3.17.4
python-dotenv>=1.0.0

# Evaluation
ragas>=0.1.0
datasets>=2.16.0

# Pattern matching
pyahocorasick>=2.0.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Utilities
tqdm>=4.66.0